pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "brainrotgen"
version = "0.1.0"
description = "Simple FastAPI boilerplate with SQLite and SQLAlchemy"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }, { include = "backend" }, { include = "db" }]
package-mode = false

[tool.poetry.dependencies]
fastapi = "^0.135.3"
pydantic-settings = "^2.6.0"
python = "^3.11"
sqlalchemy = "^2.0.35"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
gtts = "^2.5.4"
pydub = "^0.25.1"
elevenlabs = "^2.40.0"
streamlit = "^1.55.0"
requests = "^2.33.0"
starlette = "0.49.1"
pygments = "^2.20.0"

[tool.poetry.group.dev.dependencies]
black = "^26.3.1"
pytest = "^8.3.3"
flake8 = "^7.1.1"
radon = "^6.0.1"
pytest-cov = "^5.0.0"
pip-audit = "^2.8.0"
pre-commit = "^4.5.1"
isort = "^8.0.1"

[tool.black]
line-length = 80
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.venv
  | build
  | dist
)/
'''
